+++ SecretSanta/SecretSanta.Data.Tests/UserTests.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SecretSanta.Business;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -7,28 +6,27 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 
-namespace SecretSanta.Business.Tests
+namespace SecretSanta.Data.Tests
 {
     [TestClass()]
-    public class GiftTests
+    public class UserTests
 
     {
        
-        //
         [DataTestMethod]
         //correct data
-        [DataRow(1, "FirstName", "LastName", 1, "Title", "Description", "Url",false)]
+        [DataRow(1, "FirstName", "LastName",false)]
         
-        public void All_Gift_Properties_Are_Set_Correctly(int id, string firstName, string lastName, int giftId, string title, string description, string url, bool assert)
+        public void All_UserTests_Properties_Are_Set_Correctly(int id, string firstName, string lastName, bool assert)
         {
             //arrange
             User sampleUser = new User(id,firstName,lastName);
-            Gift sampleGift = new Gift(giftId,title,description,url,sampleUser);
-            IEnumerable<PropertyInfo> giftProperties = sampleGift.GetType().GetProperties();
+           
+            IEnumerable<PropertyInfo> userProperties = sampleUser.GetType().GetProperties();
 
                //Act
-                bool GiftPropertiesFilledIncorrectly = giftProperties
-                .Select(propertyInfo => { return (value: propertyInfo.GetValue(sampleGift)!, propertyInfo); })
+                bool GiftPropertiesFilledIncorrectly = userProperties
+                .Select(propertyInfo => { return (value: propertyInfo.GetValue(sampleUser)!, propertyInfo); })
                 .Any((valueAndProperty) =>
                 {
                     //not concerned about empty string values here.
@@ -56,19 +54,16 @@ namespace SecretSanta.Business.Tests
 
         [DataTestMethod]
         [ExpectedException(typeof(ArgumentNullException))]
+        //firstName null
+        [DataRow(1, null, "LastName")]
+        //lastName null
+        [DataRow(1, "firstName", null)]
      
-        //title null
-        [DataRow( 1, null, "Description", "Url")]
-        // description null
-        [DataRow( 1, "Title", null, "Url")]
-        //url null
-        [DataRow( 1, "Title", "Description", null)]
-
-        public void All_Gift_Properties_Are_Filled_Correctly_No_Nulls_User_And_Gift_Constructors_Throw_Exceptions(int giftId, string title, string description, string url)
+        public void All_User_Properties_Are_Filled_Correctly_No_Nulls_User_And_User_Constructors_Throw_Exceptions(int id, string firstName, string lastName)
         {
-            var sampleUser = new User(1, "string", "string");
+            User sampleUser = new User(id, firstName, lastName);
 
-            Gift sampleGift = new Gift(giftId, title, description, url, sampleUser);
+           
 
 
         }
