+++ SecretSanta/src/SecretSanta.Business/SecretSanta.Business.csproj
@@ -11,8 +11,4 @@
     <Compile Include="..\..\GlobalSuppression.cs" Link="GlobalSuppression.cs" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Compile Include="..\GlobalSuppressions.cs" Link="GlobalSuppressions.cs" />
-  </ItemGroup>
-
 </Project>
