+++ SecretSanta/src/SecretSanta.Data/Gift.cs
@@ -6,7 +6,7 @@ namespace SecretSanta.Data
 {
     public class Gift : FingerPrintEntityBase
     {
-        public int Id { get; }
+       
         public string Title { get => _Title; set => _Title = value ?? throw new ArgumentNullException(nameof(Title)); }
         private string _Title = string.Empty;
         public string Description { get => _Description; set => _Description = value ?? throw new ArgumentNullException(nameof(Description)); }
