+++ SecretSanta/src/SecretSanta.Data/UserGroup.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SecretSanta.Data
+{
+    public class UserGroup
+    {
+        public int UserId { get; set; }
+        public int User { get; set; }
+        public int GroupId { get; set; }
+        public int Group { get; set; }
+    }
+}
